check transactionID

+++ bitcoin.c
@@ -50,7 +50,7 @@ int main(int argc, char **argv){
     else if (strcmp(argv[i], "-b") == 0)
     {
       bucketSize = atoi(argv[i + 1]);
-      if (bucketSize < sizeof(bucket))
+      if (bucketSize < (int)sizeof(bucket))
       {
         printf("bucketSize should be more than %lu bytes\n", sizeof(bucket));
         exit(2);
